requirements.txt
streamlit
google
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
paymanai
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai